refactor: Extract TopkColumnMatcher from ContrastiveLearningAPI
This change will allows spliting the column embedding phase
from the selection of top-k most similar embeddings. Besides
allowing the introduction of alternative embedding models in
the future, this will be allow the implementation of alternative
TopkColumnMatcher's that instead of selecting the top-k based on
embedding similarities, will return the results based on
interactions of the user through with automatic recomendations
through a Jupyter notebook and visualizations.