bugfix: Support repeated source columns in match_values()
This commit fixes an edge case that breaks the match_values()
function when the input contains repeated source columns.
Previously, the matches were computed for only one of the
repeated mappings. This fix allows the output of top_matches(),
which contains repeated source values, to be passed as a
parameter of the match_values() function.